[state-prune] Refactor code

+++ incognito.go
@@ -119,7 +119,7 @@ func mainMaster(serverChan chan<- *Server) error {
 		panic(err)
 	}
 
-	p := pruner.NewPrunerWithValue(db, make(map[byte]byte))
+	p := pruner.NewPrunerWithValue(db, make(map[byte]int))
 	p.ReadStatus()
 	//check if prune flag is available
 	if config.Config().StatePrune {
